refactor: split userDetails from userEntity
